requirements.txt
aiogram>=3.2.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
aiosqlite>=0.19.0
pytz>=2023.3
aiohttp>=3.9.0